requirements.txt
Flask==0.10.0
Flask-Admin==1.4.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib=1.6.5
python-dateutil==2.5.2
six=1.10.0
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
gunicorn==19.1.1
psycopg2==2.5.4
WTForms==2.1
pbr==1.9.1
Werkzeug==0.11.5
Tempita==0.5.2
WTForms-Compontents==0.10.0
pycrypto==2.6.1
bcrypt==2.0.0
